requirements.txt
asgiref==3.4.1
beautifulsoup4==4.9.3
certifi==2021.5.30
charset-normalizer==2.0.4
click==8.0.1
Django==3.2.6
django-ckeditor==6.1.0
django-js-asset==1.2.2
django-modeltranslation==0.15.1
django-widget-tweaks==1.4.8
get-text==0.1.2
idna==3.2
importlib-metadata==4.6.4
Jinja2==3.0.1
MarkupSafe==2.0.1
Pillow==8.3.1
pkg_resources==0.0.0
python-gettext==4.0
pytz==2021.1
PyYAML==5.4.1
requests==2.26.0
six==1.16.0
soupsieve==2.2.1
sqlparse==0.4.1
typing-extensions==3.10.0.0
urllib3==1.26.6
zipp==3.5.0
